+++ example/three/index.js
@@ -78,7 +78,7 @@ animate();
 
 function reinstantiateTiles() {
 
-	const url = hashUrl || '../data/tileset.json';
+	const url = hashUrl || '../../data/tileset.json';
 
 	if ( tiles ) {